# scripts/_nexus_openapi.yaml
...
openapi: 3.0.3
info:
  title: Nexus Mods API
  version: 3.0.0
  description: "## Overview\nBuild tools and integrations for the Nexus Mods platform
    using our API.\n\n* Access mod data, user information, and platform features to
    power your own applications, services, and workflows.\n* Use our OpenAPI specification
    to explore and integrate the API.\n\n### I want to\n#### upload a mod\n* **Using
    GitHub to manage your mod workflow?** Try our [GitHub action](https://github.com/marketplace/actions/upload-to-nexus-mods)\n
    \ and [sample workflow](https://github.com/Nexus-Mods/API-Example)!\n* Or get
    started at [creating a new upload session](#tag/uploads/operation/createUpload).\n\n###
    Legacy APIs    \nLegacy integrations may still be using our [v1 API](https://app.swaggerhub.com/apis-docs/NexusMods/nexus-mods_public_api_params_in_form_data/1.0#/).\n\n###
    API Stability\n\nEndpoints are marked with badges indicating their stability level:\n\n|
    Badge | Meaning |\n| ----- | ------- |\n| *(none)* | **Stable** - Production ready.
    No breaking changes. |\n| Beta | Feature complete but may have minor changes before
    stabilisation. |\n| Experimental | May change significantly or be removed. Not
    recommended for production. |\n\nStable endpoints may receive additive enhancements
    (new optional fields/parameters)\nthat won't break existing integrations.\n\n###
...
    it will be marked as\n`deprecated` with a minimum **90-day deprecation period**
...
servers:
...
security:
- ApiKeyAuth: []
- BearerJwtAuth: []
paths:
  "/collections/{id}":
    patch:
      tags:
      - collections
      x-badges:
      - name: Experimental
        color: orange
      summary: Edit collection
      description: 'Update the core details of a collection such as the name, summary,
        description, and category.

        '
      operationId: editCollection
      parameters:
      - name: id
        in: path
        required: true
        description: The unique identifier for the collection.
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EditCollectionRequest"
      responses:
        '204':
          description: Successfully updated the collection.
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '404':
          description: The collection was not found.
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '422':
          description: The request body failed validation.
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
...
  "/collections":
    post:
      tags:
      - collections
      x-badges:
      - name: Experimental
        color: orange
      summary: Create a collection
      description: 'Creates a collection and claims an upload using the upload_id
        from a finalised [upload session](#tag/uploads/operation/createUpload).

        '
      operationId: createCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateCollectionRequest"
      responses:
        '200':
          description: Upload
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    "$ref": "#/components/schemas/CreateCollectionSuccess"
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '422':
          description: Unprocessable Entity
          content:
            application/problem+json:
              schema:
                anyOf:
                - "$ref": "#/components/schemas/ProblemDetails"
                - "$ref": "#/components/schemas/ValidationProblem"
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: apikey
      description: Get your personal API key from https://www.nexusmods.com/settings/api-keys
    BearerJwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'A signed JWT sent as `Authorization: Bearer <token>`.'
  schemas:
    ProblemDetails:
      type: object
      description: An RFC 9457 Problem Details object.
      required:
      - type
      - title
      - status
      - detail
      - instance
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code.
          example: 404
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of
            the problem.
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of
            the problem.
    ValidationProblem:
      description: RFC 9457 Problem Details extended with validation errors.
      allOf:
      - "$ref": "#/components/schemas/ProblemDetails"
      - type: object
        required:
        - errors
        properties:
          errors:
            type: array
            description: List of validation errors.
            items:
              "$ref": "#/components/schemas/ValidationProblemItem"
    ValidationProblemItem:
      type: object
      description: A single validation error with a JSON pointer to the problematic
        field.
      required:
      - detail
      - pointer
      properties:
        detail:
          type: string
          description: Human-readable description of the validation error.
        pointer:
          type: string
          description: A JSON Pointer (RFC 6901) to the field that caused the error.
          example: "/collection_data/collection_manifest/mods/0/source/mod_id"
...
    EditCollectionRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the collection.
          minLength: 3
          maxLength: 36
        summary:
          type: string
          description: Collection summary.
          maxLength: 255
        description:
          type: string
          description: Description of the collection.
        category_id:
          type: integer
          nullable: true
          description: ID of the parent category.
...
    CreateCollectionRequest:
      type: object
      required:
      - upload_id
      - collection_data
      properties:
        upload_id:
          type: string
          format: uuid
          nullable: false
          description: The unique identifier for the upload to claim against this
            new collection.
        collection_data:
          "$ref": "#/components/schemas/CollectionPayload"
    CreateCollectionSuccess:
      type: object
      required:
      - id
      - slug
      - revision_id
      - revision_number
      - revision_status
      properties:
        id:
          type: string
          description: The unique identifier for the collection.
        slug:
          type: string
          description: The slug for the collection.
        revision_id:
          type: string
          description: The unique identifier for the initial collection revision.
        revision_number:
          type: integer
          description: The revision number.
        revision_status:
          "$ref": "#/components/schemas/RevisionStatus"
...
    CollectionPayload:
      type: object
      description: The data payload used to create a collection revision.
      required:
      - adult_content
      - collection_manifest
      - collection_schema_id
      properties:
        adult_content:
          type: boolean
          description: Whether the collection includes adult content.
        collection_manifest:
          "$ref": "#/components/schemas/CollectionManifest"
        collection_schema_id:
          type: integer
          description: Collection schema id.
    CollectionManifest:
      type: object
      description: The JSON manifest that defines a collection.
      required:
      - info
      - mods
      properties:
        info:
          "$ref": "#/components/schemas/CollectionManifestInfo"
        mods:
          type: array
          items:
            "$ref": "#/components/schemas/CollectionManifestMod"
          description: List of mod resources for the manifest.
    CollectionManifestInfo:
      type: object
      description: The info section of the JSON manifest.
      required:
      - author
      - name
      - domain_name
      properties:
        author:
          type: string
          description: The collection author's name.
        author_url:
          type: string
          nullable: true
          description: The url of the author's profile.
        name:
          type: string
          description: The name of the collection.
        description:
          type: string
          nullable: true
          description: A description of the collection.
        summary:
          type: string
          nullable: true
          description: A short summary of the collection.
        domain_name:
          type: string
          description: The domain name of the game.
        game_versions:
          type: array
          nullable: true
          items:
            type: string
          description: A list of game versions that this revision has been tested
            with.
    CollectionManifestMod:
      type: object
      description: Defines a mod to be used in a collection as part of the manifest.
      required:
      - name
      - version
      - optional
      - domain_name
      - source
      properties:
        name:
          type: string
          description: The name of the mod.
        version:
          type: string
          description: The mod version.
        optional:
          type: boolean
          description: Whether the mod is required for this collection.
        domain_name:
          type: string
          description: The domain name of the game for the mod.
        source:
          "$ref": "#/components/schemas/CollectionManifestModSource"
        author:
          type: string
          nullable: true
          description: The name of the mod author.
    CollectionManifestModSource:
      type: object
      description: Source information for a mod (nexus or other) as part of the manifest.
      required:
      - type
      properties:
        type:
          "$ref": "#/components/schemas/ModSource"
        mod_id:
          type: string
          description: The identifier of the mod source for this collection manifest.
        file_id:
          type: string
          description: The identifier of the file associated with the mod in this
            collection manifest.
        md5:
          type: string
          nullable: true
          description: An MD5 hash of the file for verification.
        file_size:
          type: integer
          nullable: true
          description: The file size in kb.
        update_policy:
          "$ref": "#/components/schemas/UpdatePolicy"
        logical_filename:
          type: string
          nullable: true
          description: Logical file name of the mod resource.
        file_expression:
          type: string
          nullable: true
          description: File expression of the mod resource.
        url:
          type: string
          nullable: true
          description: The direct url of the file.
        adult_content:
          type: boolean
          nullable: true
          description: Does the mod include adult content.
    ModSource:
      type: string
      description: Used to define the source of a mod. Useful for informing Collections
        how to retrieve mods.
      enum:
      - nexus
      - direct
      - browse
      - manual
      - bundle
    RevisionStatus:
      type: string
      description: The status of a collection revision.
      enum:
      - draft
      - published
      - retracted
    UpdatePolicy:
      type: string
      description: Available update policies of a file resource
      enum:
      - exact
      - latest
      - prefer