# src/cyberpunk_mod_manager/api/app.py
# -*- coding: utf-8 -*-
"""FastAPI 应用入口。

提供两类接口：
1. REST 路由：模组 CRUD、安装/卸载、卸载计划查询
2. Agent 路由：通过 AgentScope Agent 处理自然语言请求（给出 mod_id 自动安装）

启动：uvicorn cyberpunk_mod_manager.api.app:app --reload
"""
from __future__ import annotations

from contextlib import asynccontextmanager
from pathlib import Path

from fastapi import FastAPI, HTTPException
from fastapi.responses import FileResponse
from fastapi.staticfiles import StaticFiles

from ..config import ConfigError, config
from ..storage.db import init_db
from .routes_mods import router as mods_router
from .routes_agent import router as agent_router
from .routes_config import router as config_router
from .routes_collections import router as collections_router

WEB_DIR = Path(__file__).resolve().parent.parent / "web"


@asynccontextmanager
async def lifespan(app: FastAPI):
    """应用生命周期：启动时确保数据目录存在并初始化数据库。"""
    if config.has_data_dir:
        try:
            config.ensure_dirs()
            init_db()
        except ConfigError:
            pass
    yield


app = FastAPI(title="Cyberpunk 2077 Mod Manager", lifespan=lifespan)

app.include_router(config_router, prefix="/api/config", tags=["config"])
app.include_router(collections_router, prefix="/api/collections", tags=["collections"])
app.include_router(mods_router, prefix="/api/mods", tags=["mods"])
app.include_router(agent_router, prefix="/api/agent", tags=["agent"])


@app.get("/api/health")
async def health() -> dict:
    """健康检查端点。"""
    nexus_valid = False
    nexus_user: dict = {}
    if config.nexus_api_key:
        from ..nexus.client import NexusClient

        try:
            async with NexusClient() as client:
                nexus_valid = await client.validate_key()
                if nexus_valid:
                    profile = await client.get_user_profile()
                    if profile is not None:
                        nexus_user = profile.model_dump()
        except Exception:
            nexus_valid = False
    return {
        "status": "ok",
        "game_path": config.game_path,
        "data_dir": str(config.data_dir) if config.has_data_dir else "",
        "data_dir_configured": config.has_data_dir,
        "config_file": config.config_file,
        "nexus_configured": bool(config.nexus_api_key),
        "nexus_valid": nexus_valid,
        "nexus_user": nexus_user,
        "nexus_premium": bool(nexus_user.get("is_premium")),
        "llm_configured": bool(config.openai_api_key),
    }


# 静态前端：/assets 资源 + SPA 回退（刷新子路径仍返回 index.html）
assets_dir = WEB_DIR / "assets"
if assets_dir.is_dir():
    app.mount("/assets", StaticFiles(directory=str(assets_dir)), name="assets")


@app.get("/")
async def serve_index() -> FileResponse:
    return FileResponse(WEB_DIR / "index.html")


@app.get("/{full_path:path}")
async def serve_spa(full_path: str) -> FileResponse:
    if full_path.startswith("api/"):
        raise HTTPException(404)
    if full_path.startswith("assets/"):
        raise HTTPException(404)
    candidate = WEB_DIR / full_path
    if candidate.is_file():
        return FileResponse(candidate)
    return FileResponse(WEB_DIR / "index.html")